requirements.txt
openai==1.3.0
sentence-transformers>=2.5.0
chromadb==0.4.18
pdfplumber==0.10.3
streamlit==1.28.0
python-dotenv==1.0.0